massively changed pause functionality, pause works now
