--- Assets/Scripts/Interfaces/Pauseable.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ca214422f58b3394f868d7d1cca6b2aa
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
