+++ Assets/Scripts/Level 2/Bookcase.cs
@@ -1,8 +1,6 @@
-﻿using MoreMountains.FeedbacksForThirdParty;
-using MoreMountains.InventoryEngine;
-using UnityEngine;
+﻿using UnityEngine;
 
-public class Bookcase : Pauseable
+public class Bookcase : MonoBehaviour
 {
     [Header("Will check for this item to repair bookshelf")]
     public Item CheckItem;
@@ -32,9 +30,8 @@ public class Bookcase : Pauseable
     [Header("Time it takes to trigger the earthquake after the bookcase is secured")]
     public float TriggerTime = 5f;
     
-    private new void Start()
+    private void Start()
     {
-        base.Start();
         _interact = GetComponent<InteractWithObject>();
         _inventory = GameObject.FindWithTag("MainInventory").GetComponent<InventoryHelper>();
         if (CheckItem == null) Debug.LogError("No item to check has been specified");
@@ -45,20 +42,6 @@ public class Bookcase : Pauseable
         fallCollider.enabled = false;
 
         //QuakeManager.Instance.OnQuake.AddListener(Fall);
-    }
-
-    public override void Pause()
-    {
-        vel = rb.velocity;
-        ang = rb.angularVelocity;
-        rb.isKinematic = true;
-    }
-
-    public override void UnPause()
-    {
-        rb.isKinematic = false;
-        rb.velocity = vel;
-        rb.angularVelocity = ang;
     }
 
     public void UpdateState()
@@ -130,11 +113,5 @@ public class Bookcase : Pauseable
             Debug.Log("Player Hit");
             DeathManager.Instance.PlayerDeath("Your bookcase crushed you :(");            
         }
-        
     }
-
-
-
-
-
 }
