+++ Assets/Scripts/Non-Specific/Clobberer.cs
@@ -4,15 +4,9 @@ using UnityEngine;
 /// <summary>
 ///  When the player enters a collider with this script they will be killed
 /// </summary>
-public class Clobberer : Pauseable
+public class Clobberer : MonoBehaviour
 {
     public bool enabled;
-    private Rigidbody rb;
-
-    private void Awake()
-    {
-        rb = GetComponent<Rigidbody>();
-    }
 
     private void OnCollisionEnter(Collision other)
     {
@@ -20,15 +14,5 @@ public class Clobberer : Pauseable
         {
             DeathManager.Instance.PlayerDeath("You were hit by a door!");
         }
-    }
-
-    public override void Pause()
-    {
-        rb.isKinematic = true;
-    }
-
-    public override void UnPause()
-    {
-        rb.isKinematic = false;
     }
 }
