+++ Assets/Scripts/Non-Specific/FallingObject.cs
@@ -1,17 +1,15 @@
-﻿using System.Collections;
+﻿using System;
+using System.Collections;
 using UnityEngine;
 
 /// <summary>
 /// When a piece of furniture falls on the player from the quake it kills them.
 /// after it hits the ground ~1.5f it will disable the ability to kill
 /// </summary>
-public class FallingObject : Pauseable
+public class FallingObject : MonoBehaviour
 {
     private bool isEnabled = true;
     private Rigidbody rb;
-    
-    private Vector3 _velocity;
-    private Vector3 _angular;
 
     private void Awake()
     {
@@ -36,20 +34,5 @@ public class FallingObject : Pauseable
         {
             DeathManager.Instance.PlayerDeath("Crushed by falling object "+ (QuakeManager.Instance.quakes == 0? "in earthquake":"in aftershock"));
         }
-    }
-
-    public override void Pause()
-    {
-        _velocity = rb.velocity;
-        _angular = rb.angularVelocity;
-        rb.isKinematic = true;
-    }
-
-    public override void UnPause()
-    {
-        rb.isKinematic = false;
-        
-        rb.velocity = _velocity ;
-        rb.angularVelocity = _angular;
     }
 }
