+++ Assets/Scripts/Non-Specific/ObjectDropper.cs
@@ -8,7 +8,7 @@ using UnityEngine;
 /// Drops and staggers the falling furniture, the longer the quake goes, the faster stuff falls, at the very end
 /// we drop an object directly on the players head if they still haven't gotten under the table or died
 /// </summary>
-public class ObjectDropper: Pauseable
+public class ObjectDropper: MonoBehaviour
 {
     public GameObject[] falling_objects;
     
@@ -24,25 +24,13 @@ public class ObjectDropper: Pauseable
     private float fallRate;
     
     private bool dropping;
-    private bool paused;
     private Vector3 playerTransform;
     private int i = 0;
     
-    private new void Start()
+    private void Start()
     {
-        base.Start();
         player = GameObject.FindWithTag("Player");
         fallRate = TotalDropTime / (falling_objects.Length + 1);
-    }
-
-    public override void Pause()
-    {
-        paused = true;
-    }
-
-    public override void UnPause()
-    {
-        paused = false;
     }
 
     public void Drop()
@@ -57,10 +45,7 @@ public class ObjectDropper: Pauseable
 
         foreach (var obj in falling_objects)
         {
-            // if paused just wait
-            while (paused) yield return new WaitForEndOfFrame();
-            
-                if (!obj.activeInHierarchy)
+            if (!obj.activeInHierarchy)
             {
                 obj.SetActive(true);
                 if(fallRate > 0) yield return new WaitForSeconds(fallRate);
