+++ Assets/Scripts/Non-Specific/PauseManager.cs
@@ -1,12 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using UnityEngine;
+﻿using UnityEngine;
 
 public class PauseManager : MonoBehaviour
 {
     public static PauseManager Instance;
-
-    private List<Pauseable> pauseAgents;
+    
     private bool paused;
     
     private void Awake()
@@ -14,48 +11,18 @@ public class PauseManager : MonoBehaviour
         if (Instance == null) Instance = this;
         else Destroy(this);
         
-        pauseAgents = new List<Pauseable>();
     }
-
-    public void Register(Pauseable agent)
-    {
-        pauseAgents.Add(agent);
-    }
-
     public bool Pause(bool val)
     {
-        if (val && !paused)
+        if (val)
         {
-            Pause();
+            Time.timeScale = 0;
+            paused = true;
             return true;
         }
         
-        if (paused)
-        {
-            UnPause();
-        }
+        Time.timeScale = 1f;
+        paused = false;
         return false;
     }
-
-    public void Pause()
-    {
-        if (paused) return;
-        paused = true;
-        foreach (Pauseable pauseAgent in pauseAgents)
-        {
-            pauseAgent.Pause();
-        }
-    }
-
-    public void UnPause()
-    {
-        if (!paused) return;
-        paused = false;
-        
-        foreach (Pauseable pauseAgent in pauseAgents)
-        {
-            pauseAgent.UnPause();
-        }
-    }
-    
 }
