--- Assets/Scripts/Non-Specific/PlayerPauser.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 55d4d097eae04b458bd1b555eb774848
-timeCreated: 1582667452
